requirements.txt
asgiref==3.5.2
astroid==2.11.2
attrs==22.1.0
autopep8==1.6.0
certifi==2022.6.15.1
charset-normalizer==2.1.1
colorama==0.4.4
dill==0.3.4
distlib==0.3.6
Django==4.1.1
django-cors-headers==3.13.0
django-grappelli==3.0.3
django-suit @ https://github.com/darklow/django-suit/tarball/v2
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
filelock==3.8.0
idna==3.3
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mysql-connector-python==8.0.33
mysqlclient==2.1.1
numpy==1.23.2
packaging==21.3
Pillow==9.2.0
pipenv==2022.9.8
platformdirs==2.5.1
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyfiglet==0.8.post1
pylint==2.13.2
pyparsing==3.0.9
pytest==7.1.3
pytz==2022.6
regex==2022.9.13
requests==2.28.1
sqlparse==0.4.2
termcolor==1.1.0
toml==0.10.2
tomli==2.0.1
tzdata==2022.2
urllib3==1.26.12
virtualenv==20.16.5
virtualenv-clone==0.5.7
windows-curses==2.3.0
wrapt==1.14.0
